// frontend/src/dto/TraceState.d.ts
export interface PrimitiveVal {
  readonly kind: 'PrimitiveVal'
  title: string | undefined // value displayed in a title tag (visible on hover)
  vizValue: string // value displayed as text, which might differ from original value e.g. when truncating float mantissa
  readonly primitiveValue: string
}

export interface ReferenceVal {
  readonly kind: 'ReferenceVal'
  readonly reference: number
}

export interface NullVal {
  readonly kind: 'NullVal'
}

export type Val = PrimitiveVal | ReferenceVal | NullVal

interface IVar { // interface to include these fields in all members of the union type Var
  readonly kind: 'Var' // the backend does not distinguish static vars and local vars, everything therefore has kind Var
  readonly name: string
  readonly type: string
  readonly value: Val
  changed: boolean
  heapObjectId?: number
}

export interface LocalVar extends IVar {
  class: string
  method: string
}

export interface StaticVar extends IVar {
  class: string
}
// ...
export interface HeapArrayElementVar {
  readonly kind: 'HeapArrayElementVar'
  readonly arrayId: number
  readonly type: string
  readonly value: Val
  readonly index: number
  changed: boolean
}

export type Var = LocalVar | StaticVar

interface IHeapItem {
  readonly id: number
  readonly type: string
  readonly faked: boolean
}

export interface HeapArray extends IHeapItem {
  readonly kind: 'HeapArray'
  readonly elements: HeapArrayElementVar[]
}

export interface HeapString extends IHeapItem {
  readonly kind: 'HeapString'
  string: string
  vizString: string
  readonly charArr: Var
}

export interface HeapObject extends IHeapItem {
  readonly kind: 'HeapObject'
  readonly fields: Var[]
  /** True if this object's fields were collected for specialised visualisations only and must not be shown in the Memory View. */
  readonly detailedFieldsOnly?: boolean
}